Fix potential segfaults in mythtranscode by wrapping a couple places that use m_proginfo with checks to make sure m_proginfo is not null.  Second part of patch by Brendan Cully on the -dev mailing list, first part by me.
git-svn-id: http://svn.mythtv.org/svn/trunk@8921 7dbf422c-18fa-0310-86e9-fd20926502f2

+++ mythtv/programs/mythtranscode/transcode.cpp
@@ -774,7 +774,7 @@ int Transcode::TranscodeFile(char *inputname, char *outputname,
         }
         if (QDateTime::currentDateTime() > curtime) 
         {
-            if (honorCutList && 
+            if (honorCutList && m_proginfo &&
                 m_proginfo->CheckMarkupFlag(MARK_UPDATED_CUT))
             {
                 unlink(outputname);
@@ -807,9 +807,12 @@ int Transcode::TranscodeFile(char *inputname, char *outputname,
 
     if (! fifow)
     {
-        m_proginfo->ClearPositionMap(MARK_KEYFRAME);
-        m_proginfo->ClearPositionMap(MARK_GOP_START);
-        m_proginfo->ClearPositionMap(MARK_GOP_BYFRAME);
+        if (m_proginfo)
+        {
+            m_proginfo->ClearPositionMap(MARK_KEYFRAME);
+            m_proginfo->ClearPositionMap(MARK_GOP_START);
+            m_proginfo->ClearPositionMap(MARK_GOP_BYFRAME);
+        }
 
         nvr->WriteSeekTable();
         if (!kfa_table->isEmpty())
